Move Spec into benchmark registration
